PDDF: Get pddf i2c package building under Debian Trixie
This is a quick attempt to modernize the Debian Packaging of the
pddf/i2c package to allow it to build under Debian Trixie. Specifically,
upgrade the compat level to a modern level (12) and update the rules
file to adhere to the modern standards.

The driver files needed to be modified to compile on both 6.1 and 6.12
kernels. The probe method for i2c only takes 1 argument as of Linux 6.2
and there was a transitional "probe_new" method created before that to
help with the porting. We now conditionally compile with probe_new and
probe depending on the kernel version targetted.